Ensure valid path and allow for file suffixes
Use method to create file path instead of concatenation to ensure valid path string. The current implementation requires a / as last element of the path but does not check for it. Added suffix variable to allow for file name pattern. This will retrieve only files with a given suffix from the provided directory. Passing the suffix through the cli needs to be implemented.